To support sensor hub function.
SensorHub:
motion sensor: MPU-6050.
mag sensor: QMC5883L.
PHT sensors: AHT20 and BMP280.

+++ software/arduino-platformio-vscode/lib/App/Common/HAL/HAL_Def.h
@@ -43,10 +43,10 @@ typedef struct
 /* MAG */
 typedef struct
 {
-    float dir;
-    float x;
-    float y;
-    float z;
+    int dir;
+    int x;
+    int y;
+    int z;
 } MAG_Info_t;
 
 /* IMU */
